Bugfix/BSR-1549 added an option(--full-sync-on-fail) to allow promotion to succeed even if the initial file system check fails, and instead proceed with a full sync.

Approved-by: 최송아
Approved-by: SungEun Kim

+++ bsr/bsr_sender.c
@@ -2853,7 +2853,7 @@ int w_e_end_ov_reply(struct bsr_work *w, int cancel)
 		bsr_advance_rs_marks(peer_device, peer_device->ov_left);
 		// peer needs to receive ack to execute bsr_rs_complete_io()
 		// send P_OV_RESULT for sector, set size to 0
-		err = bsr_send_ack_ex(peer_device, P_OV_RESULT, sector, peer_device->ov_last_skipped_size << 9, ID_IN_SYNC); // BSR-1553 send skipped_size to target
+		err = bsr_send_ack_ex(peer_device, P_OV_RESULT, sector, (int)(peer_device->ov_last_skipped_size << 9), ID_IN_SYNC); // BSR-1553 send skipped_size to target
 	}
 
 	dec_unacked(peer_device);
